Move client issuance folder into issuer/manifest
The Credential Application is part of the Credential Manifest spec so it
makes sense to be in the manifest file.
