--- packages/verity/lib/client/index.ts
@@ -1,1 +1,0 @@
-export * from "./issuance"
